fix(blind-bidding): allow for resubmitting higher bids
Generally, a player should go to the highest bidder. If Owner A submits a $0 bid for a player, then the next day Owner B submits a $10 bid, the $0 bid loses. If Owner B is paranoid and wants to increase the bid to $20, the owner should be able to cancel and submit a new bid for $20 without Owner A winning the player for $0 as long as the cancel/submit for $20 happened on the same day.

Closes #94